Rearrange ProxyForm related code
